Added level generator script
This update adds the ability to generate the level grid by designating the
level size (x * y) and the outline of each grid block. Notably, an
Editor folder has been added with a level editor script so that the
level can be adjusted in real time, which will be helpful while designing levels.