Removed the ability for the player to spawn camp
Prior towards this push, the player could find a good "baracade" on the level
and just wait for the enemies to funnel towards them. To fix this, if the
player doesn't move from their current coordinates within a setable threshold,
the enemies will begin to spawn on the player. This makes it so the player must
move from their camped position.